+++ apps/portals/standards/tsconfig.json
@@ -21,7 +21,7 @@
       "vite-config/*": ["../../../packages/vite-config/src/*"]
     }
   },
-  "include": ["**/*.tsx", "**/*.ts"],
+  "include": ["src/**/*.tsx", "src/**/*.ts"],
   "files": ["src/types.d.ts"],
   "references": [
     { "path": "../../synapse-portal-framework" },
