+++ packages/synapse-react-client/src/stories/ColorPaletteInspector.stories.ts
