+++ packages/synapse-react-client/src/stories/stubs/Alert/MuiAlert.stories.tsx
