+++ packages/synapse-react-client/src/stories/stubs/Alert/MuiAlert.tsx
