+++ packages/synapse-react-client/src/stories/stubs/Button/MuiButton.stories.tsx
