+++ packages/synapse-react-client/src/stories/stubs/Button/MuiButton.tsx
