+++ packages/synapse-react-client/src/stories/stubs/Dialog/MuiDialog.stories.tsx
