+++ packages/synapse-react-client/src/stories/stubs/Dialog/MuiDialog.tsx
@@ -10,8 +10,8 @@ import {
   Stack,
 } from '@mui/material'
 import { ReactNode } from 'react'
-import { HelpPopover } from '../../../src/components/HelpPopover/HelpPopover'
-import IconSvg from '../../../src/components/IconSvg/IconSvg'
+import { HelpPopover } from '../../../components/HelpPopover/HelpPopover'
+import IconSvg from '../../../components/IconSvg/IconSvg'
 
 export interface DialogProps extends Omit<MuiDialogProps, 'content'> {
   title: string
