+++ packages/synapse-react-client/src/stories/stubs/Tooltip/MuiTooltip.tsx
