+++ packages/synapse-react-client/src/stories/stubs/Typography.stories.tsx
