+++ packages/synapse-react-client/src/testutils/MockSvg.js
