+++ packages/synapse-types/tsconfig.json
@@ -3,7 +3,7 @@
   "compilerOptions": {
     "composite": true,
     "declaration": true,
-    "rootDir": "./src",
+    "rootDir": "src",
     "outDir": "./dist",
     "module": "esnext",
     "target": "esnext",
